created RowLabel and RowValue components to further reduce complexity of Row component and too later support AI model configurations
